% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HC.R
\name{theme_HC()}
\alias{theme_HC()}
\alias{theme_HC}
\title{A black High Contrast theme for ggplot2}
\usage{
theme_HC()
}
\description{
This function applies a  black high contrast theme to ggplot2 for use by people with substantially impaired vision .
}
\examples{
library(ggplot2)
ggplot(diamonds, aes(x = carat, y = price)) +
geom_point(colour = "Lightgreen", size =2) +
theme_HC()
}
\keyword{dark}
