% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HC_blue.R
\name{theme_HC_blue()}
\alias{theme_HC_blue()}
\alias{theme_HC_blue}
\title{Blue High Contrast theme for ggplot2}
\usage{
theme_HC_blue()
}
\description{
This function applies a  blue high contrast theme to ggplot2 for use by people with substantially impaired vision .
}
\examples{
library(ggplot2)
ggplot(diamonds, aes(x = carat, y = price)) +
geom_point(colour = "Lightgreen", size =2) +
theme_HC_blue()
}
\keyword{blue}
