% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HC_red.R
\name{theme_HC_red()}
\alias{theme_HC_red()}
\alias{theme_HC_red}
\title{red High Contrast theme for ggplot2}
\usage{
theme_HC_red()
}
\description{
This function applies a  red high contrast theme to ggplot2 for use by people with substantially impaired vision .
}
\examples{
library(ggplot2)
ggplot(diamonds, aes(x = carat, y = price)) +
geom_point(colour = "yellow", size =2) +
theme_HC_red()
}
\keyword{red}
